20240821: 배포

+++ src/main/java/com/finalproject/sulbao/product/controller/ProductController.java
@@ -1,5 +1,6 @@
 package com.finalproject.sulbao.product.controller;
 
+import com.finalproject.sulbao.product.model.dto.ProductComparisonDTO;
 import com.finalproject.sulbao.product.model.dto.ProductDTO;
 import com.finalproject.sulbao.product.model.entity.ProductCategory;
 import com.finalproject.sulbao.product.service.ProductService;
@@ -26,7 +27,7 @@ public class ProductController {
     @GetMapping("/list")
     public String productList(Model model, HttpSession session) {
 
-        if(session == null){
+        if(session.getAttribute("userNo") == null){
             return "redirect:/login";
         }
 
@@ -34,45 +35,72 @@ public class ProductController {
         productDTO.setUserNo((Long) session.getAttribute("userNo"));
         List<ProductDTO> productList = productService.findByUserNo(productDTO);
 
+        model.addAttribute("menu","product");
+        model.addAttribute("submenu","plist");
         model.addAttribute("product", productDTO);
         model.addAttribute("productList", productList);
-        return "product/productList";
+        return "admin/product/productList";
     }
 
-    //조회
+    // 관리자 페이지 조회
     @GetMapping("/search")
     public String productSearch(Model model, @ModelAttribute ProductDTO productDTO, HttpSession session) {
 
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
+
         productDTO.setUserNo((Long) session.getAttribute("userNo"));
         List<ProductDTO> productList = productService.findBySearchInfo(productDTO);
-
+        model.addAttribute("menu","product");
+        model.addAttribute("submenu","list");
         model.addAttribute("product", productDTO);
         model.addAttribute("productList", productList);
-        return "product/productList";
+        return "admin/product/productList";
     }
 
     //등록화면으로 이동
     @GetMapping("/detail")
-    public String productRegist(Model model) {
+    public String productRegist(Model model, HttpSession session) {
+
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
+
+        List<ProductComparisonDTO> productComparisonList = productService.findByComparison();
+        model.addAttribute("menu","product");
+        model.addAttribute("submenu","regist");
+        model.addAttribute("productComparisonList", productComparisonList);
         model.addAttribute("product", new ProductDTO());
         model.addAttribute("productCategory", ProductCategory.values());
-        return "product/productDetail";
+        return "admin/product/productDetail";
     }
 
     //수정화면으로 이동
     @GetMapping("/update/{productNo}")
-    public String productUpdate(Model model,@PathVariable(required = false,name = "productNo") Long productNo) {
+    public String productUpdate(Model model,@PathVariable(required = false,name = "productNo") Long productNo, HttpSession session) {
+
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
+
+        List<ProductComparisonDTO> productComparisonList = productService.findByComparison();
+        model.addAttribute("productComparisonList", productComparisonList);
 
         ProductDTO productDTO =  productService.findByProductNo(productNo);
         model.addAttribute("product", productDTO);
         model.addAttribute("productCategory", ProductCategory.values());
-        return "product/productUpdate";
+        return "admin/product/productUpdate";
     }
 
     //상품등록
     @PostMapping("/regist")
     public String saveProduct(@ModelAttribute ProductDTO productDTO, HttpSession session) {
 
+        log.info("productController saveProduct : {}", productDTO);
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
         productDTO.setUserNo((Long) session.getAttribute("userNo"));
         // 저장
         productService.saveProduct(productDTO);
@@ -83,6 +111,9 @@ public class ProductController {
     @PostMapping("/update")
     public String updateProduct(@ModelAttribute ProductDTO productDTO, HttpSession session) {
 
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
         productDTO.setUserNo((Long) session.getAttribute("userNo"));
         productService.updateProduct(productDTO);
 
@@ -92,7 +123,10 @@ public class ProductController {
     //상품삭제
     @DeleteMapping("/delete")
     @ResponseBody
-    public String deleteProduct(String productNoList) {
+    public String deleteProduct(String productNoList, HttpSession session) {
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
         productService.delete(productNoList);
         return "success";
     }
@@ -100,9 +134,27 @@ public class ProductController {
     //상품상태변경
     @PutMapping("/status")
     @ResponseBody
-    public String updateStatus(String productNoList, String type, String status) {
+    public String updateStatus(String productNoList, String type, String status, HttpSession session) {
+        if(session.getAttribute("userNo") == null){
+            return "redirect:/login";
+        }
+        if(productNoList == null || productNoList.isEmpty()){
+            return "fail";
+        }
         productService.updateStatus(productNoList,type,status);
         return "success";
+    }
+
+    // 사용자 페이지 상품목록
+    @GetMapping("/user/list")
+    public String userList(Model model) {
+        return "product/list";
+    }
+
+    // 사용자 페이지 상품상세조회
+    @GetMapping("/user/detail")
+    public String userDetail(Model model) {
+        return "product/detail";
     }
 
 }
